20221110 #1
불필요한 부분 SQL 정리
이미지 소스 연결 추가
참석자 검색 추가
APP용 회원정보 부분 요청사항대로 수정
회원 부분 : 이메일 발송 수정
회원 부분 : 파일 저장 수정
파일 저장 함수 추가
이메일 폼 수정
회원 가입 시 미팅 참여자에 속한 정보와 연동 처리
회원 가입 인증 키 Integer에서 varchar 로 변경

+++ src/main/java/com/eura/web/service/UserService.java
@@ -74,7 +74,7 @@ public class UserService {
 
     }
 
-    public long join(UserVO userVO){
+    public Integer join(UserVO userVO){
 
         SimpleDateFormat fm = new SimpleDateFormat ("yyyy-MM-dd HH:mm:ss");
         String now_date = fm.format(Calendar.getInstance().getTime());
@@ -85,7 +85,7 @@ public class UserService {
         userVO.setLast_pw_upd_dt(now_date);
 
         userMapper.insertUserInfo(userVO);
-        long ret_idx = userMapper.findUserIdx(userVO.getUser_id());
+        Integer ret_idx = userMapper.findUserIdx(userVO.getUser_id());
 
         return ret_idx;
     }
